Refactor the application structure; split classes into separate files
