+++ app/ruby_engine.rb
@@ -1,0 +1,18 @@
+require 'ruby_engine/opal'
+require 'ruby_engine/cruby_wasi'
+
+class RubyEngine
+  def run(source, instance)
+    raise NotImplementedError
+  end
+
+  ENGINES = [
+    Opal.new,
+    CRubyWASI.new(
+      "https://cdn.jsdelivr.net/npm/ruby-wasm-wasi@0.1.2/dist/ruby.wasm",
+      "3.2.0dev"
+    ),
+  ].each_with_object({}) do |engine, hash|
+    hash[engine.engine_id] = engine
+  end
+end
