+++ source/javascripts/application.js
@@ -1,6 +1,3 @@
 //= require try_ruby
 
-document.addEventListener("DOMContentLoaded", () => {
-  Opal.loaded(OpalLoaded || []);
-  Opal.require("try_ruby");
-})
+Opal.require("try_ruby");
